requirements.txt
flask
flask-cors
pymongo
dnspython
python-dotenv
pandas
numpy
scikit-learn
torch
torchvision
opencv-python
pillow
transformers
nltk
matplotlib
seaborn
imutils
xgboost
joblib
scipy
moviepy
ffmpeg-python
tqdm
firebase-admin
flask-login